issue #13. Implementation without UT

+++ src/python/server/symaicorecommands.py
@@ -304,25 +304,43 @@ class SymAICoreCommands(symaicommands.SymAICommands):
             raise e
 
     def do_environment(self, cuuid, data_received):
-        res = self.get_and_simplify(cuuid, data_received, symaiconfig.SymAIConfig.BASE_ENVIRONMENT.value)
-
-        fn = os.path.join(symaiconfig.SymAIConfig.BASE_TEMP.value,
-                          cuuid,
-                          symaiconfig.SymAIConfig.BASE_ENVIRONMENT.value,
-                          res["filename"])
-        try:
-            with open(fn, "w") as f:
-                f.write(res["content"])
-            self.get_logger().info(f"environment command processed. The environment saved to {fn}")
-            setattr(self, "environment", fn)
-        except Exception as e:
-            self.get_logger().error(f"environment command processing failed {str(e)}")
+        if data_received is None or len(data_received.strip()) < 1:
+            res = dict()
+            cnt = ""
+            if hasattr(self, "environment"):
+                cnt = self.get_environment()
+            res["environment"] = cnt
+            res = json.dumps(res)
+        else:
             try:
-                if os.path.exists(fn):
-                    os.remove(fn)
-            except:
-                pass
-            raise e
+                res = json.loads(data_received)
+                if "filename" not in res.keys():
+                    res["filename"] = "environment.env"
+                    data_received = json.dumps(res)
+            except Exception as e:
+                self.get_logger().error(f"environment command processing failed {str(e)}")
+                raise e
+            res = self.get_and_simplify(cuuid, data_received, symaiconfig.SymAIConfig.BASE_ENVIRONMENT.value)
+
+            fn = os.path.join(symaiconfig.SymAIConfig.BASE_TEMP.value,
+                              cuuid,
+                              symaiconfig.SymAIConfig.BASE_ENVIRONMENT.value,
+                              res["filename"])
+            try:
+                with open(fn, "w") as f:
+                    f.write(res["content"])
+                self.get_logger().info(f"environment command processed. The environment saved to {fn}")
+                setattr(self, "environment", fn)
+                res = ""
+            except Exception as e:
+                self.get_logger().error(f"environment command processing failed {str(e)}")
+                try:
+                    if os.path.exists(fn):
+                        os.remove(fn)
+                except:
+                    pass
+                raise e
+        return res
 
     def do_property(self, cuuid, data_received):
         res = self.get_and_simplify(cuuid, data_received, symaiconfig.SymAIConfig.BASE_PROPERTIES.value)
